backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
httpx==0.27.2
resend==2.4.0
python-dotenv==1.0.1
pydantic[email]==2.8.2
